app/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
Flask==2.3.2
Flask-Mail==0.9.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python
mysqlclient==2.2.0
pycparser==2.21
Werkzeug==2.3.6